+++ Packages/com.unity.inputsystem/InputSystem/Actions/InputActionState.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using System.Runtime.InteropServices;
 using Unity.Collections;
 using Unity.Collections.LowLevel.Unsafe;
+using Unity.Profiling;
 using UnityEngine.InputSystem.Controls;
 using UnityEngine.InputSystem.LowLevel;
 using UnityEngine.InputSystem.Utilities;
@@ -122,6 +123,9 @@ namespace UnityEngine.InputSystem
         private InputEventPtr m_CurrentlyProcessingThisEvent;
         private Action m_OnBeforeUpdateDelegate;
         private Action m_OnAfterUpdateDelegate;
+        private static readonly ProfilerMarker s_InputInitialActionStateCheckMarker = new ProfilerMarker(ProfilerCategory.Input, "InitialActionStateCheck");
+        private static readonly ProfilerMarker s_InputActionResolveConflictMarker = new ProfilerMarker(ProfilerCategory.Input, "InputActionResolveConflict");
+        private static readonly ProfilerMarker s_InputActionCallbackMarker = new ProfilerMarker(ProfilerCategory.Input, "InputActionCallback");
 
         /// <summary>
         /// Initialize execution state with given resolved binding information.
@@ -1257,7 +1261,7 @@ namespace UnityEngine.InputSystem
             // Remove us from the callback as the processing we're doing here is a one-time thing.
             UnhookOnBeforeUpdate();
 
-            Profiler.BeginSample("InitialActionStateCheck");
+            s_InputInitialActionStateCheckMarker.Begin();
 
             // Use current time as time of control state change.
             var time = InputState.currentTime;
@@ -1315,7 +1319,7 @@ namespace UnityEngine.InputSystem
             }
             manager.FireStateChangeNotifications();
 
-            Profiler.EndSample();
+            s_InputInitialActionStateCheckMarker.End();
         }
 
         // Called from InputManager when one of our state change monitors has fired.
@@ -1657,7 +1661,7 @@ namespace UnityEngine.InputSystem
             // Anything below here we want to avoid executing whenever we can.
             Debug.Assert(actionState->mayNeedConflictResolution);
 
-            Profiler.BeginSample("InputActionResolveConflict");
+            s_InputActionResolveConflictMarker.Begin();
 
             // We take a local copy of this value, so we can change it to use the starting control of composites
             // for simpler conflict resolution (so composites always use the same value), but still report the actually
@@ -1691,7 +1695,7 @@ namespace UnityEngine.InputSystem
             if (actionStateControlIndex == kInvalidIndex)
             {
                 actionState->magnitude = trigger.magnitude;
-                Profiler.EndSample();
+                s_InputActionResolveConflictMarker.End();
                 return false;
             }
 
@@ -1720,7 +1724,7 @@ namespace UnityEngine.InputSystem
 
                 // Keep recorded magnitude in action state up to date.
                 actionState->magnitude = trigger.magnitude;
-                Profiler.EndSample();
+                s_InputActionResolveConflictMarker.End();
                 return false;
             }
 
@@ -1734,7 +1738,7 @@ namespace UnityEngine.InputSystem
                 // actuation as we didn't have the highest actuation anyway.
                 if (!isControlCurrentlyDrivingTheAction)
                 {
-                    Profiler.EndSample();
+                    s_InputActionResolveConflictMarker.End();
                     ////REVIEW: should we *count* actuations instead? (problem is that then we have to reliably determine when a control
                     ////        first actuates; the current solution will occasionally run conflict resolution when it doesn't have to
                     ////        but won't require the extra bookkeeping)
@@ -1749,7 +1753,7 @@ namespace UnityEngine.InputSystem
                 {
                     // Keep recorded magnitude in action state up to date.
                     actionState->magnitude = trigger.magnitude;
-                    Profiler.EndSample();
+                    s_InputActionResolveConflictMarker.End();
                     return false;
                 }
 
@@ -1868,12 +1872,12 @@ namespace UnityEngine.InputSystem
                     actionState->bindingIndex = bindingWithHighestActuation;
                     actionState->magnitude = highestActuationLevel;
 
-                    Profiler.EndSample();
+                    s_InputActionResolveConflictMarker.End();
                     return false;
                 }
             }
 
-            Profiler.EndSample();
+            s_InputActionResolveConflictMarker.End();
 
             // If we're not really effecting any change on the action, ignore the control state change.
             // NOTE: We may be looking at a control here that points in a completely direction, for example, even
@@ -2520,7 +2524,7 @@ namespace UnityEngine.InputSystem
                 m_ActionIndex = actionIndex,
             };
 
-            Profiler.BeginSample("InputActionCallback");
+            s_InputActionCallbackMarker.Begin();
 
             // Global callback goes first.
             var action = context.action;
@@ -2552,7 +2556,7 @@ namespace UnityEngine.InputSystem
             // Run callbacks (if any) on action map.
             DelegateHelpers.InvokeCallbacksSafe(ref callbacksOnMap, context, callbackName, actionMap);
 
-            Profiler.EndSample();
+            s_InputActionCallbackMarker.End();
         }
 
         private object GetActionOrNoneString(ref TriggerState trigger)